type_traits: Separate type traits helpers into a new header file
This commit adds cista/type_traits.h and move is_char_array_helper and
is_string_helper into that new header file.
